refactor(arch): split query-aware widgets out of ui/ into widgets/ (#282)
ui/ had eroded into two tiers sharing one folder and one lint boundary:
true primitives (Button, Modal, …) and domain widgets that import
lib/queries / lib/outbox. Move the four data-aware components —
PictogramSheet, KidSheet, NewKidModal, OfflineIndicator — to a new
src/widgets/ tier (shared, query-aware, feature-agnostic) and tighten
the boundaries:

- ui/ can no longer import @/lib/queries, @/lib/outbox, or @/widgets
- widgets/ joins the no-app-imports, no-feature-imports, and
  supabase.storage/auth boundary blocks
- shared layers now also explicitly forbid importing features/

KidSwitcher (listed in #282) no longer imports queries and KidModeGate
only uses lib/pin, so both stay in ui/. Mechanical move, no behavior
change; README architecture diagram and directory map updated.
