+++ src/app/routes/ParentHomeRoute.tsx
@@ -9,7 +9,7 @@ import { getLastBoard, hasAutoLaunched, kidPathFor, markAutoLaunched } from '@/l
 import { useBoards, useCreateBoard } from '@/lib/queries/boards';
 import { useActiveKid } from '@/lib/queries/kids';
 import { accentForIndex } from '@/theme/tokens';
-import { NewKidModal } from '@/ui/NewKidModal/NewKidModal';
+import { NewKidModal } from '@/widgets/NewKidModal/NewKidModal';
 
 export const ParentHomeRoute = (): JSX.Element => {
   const navigate = useNavigate();
