+++ src/widgets/NewKidModal/NewKidModal.module.css
