+++ src/widgets/NewKidModal/NewKidModal.tsx
