+++ src/widgets/OfflineIndicator/OfflineIndicator.test.tsx
