+++ src/widgets/OfflineIndicator/OfflineIndicator.tsx
