+++ src/widgets/PictogramSheet/PictogramSheet.module.css
