+++ src/widgets/PictogramSheet/PictogramSheet.test.tsx
