+++ src/widgets/PictogramSheet/PictogramSheet.tsx
